correção dos erros

+++ Exc3.java
@@ -4,7 +4,7 @@ public class Exc3 {
     public static void main(String[] args) {
         Scanner numeros = new Scanner(System.in);
         System.out.println("Informe o salário");
-        float salario = numeros.nextInt();
+        float salario = numeros.nextFloat();
         System.out.println("Informe o percentual de aumento");
         float aumento = numeros.nextFloat();
